🎨 디자인시스템 폴더 바텀 시트 UI 수정 #21
